PAGE COPAGETEC REPOSITORY ELECTRICAL CIRCUITS
Digital repositories can be described as a bank for general documents. This means that it is an archive where the generated digital resources can be stored, so that other interested people can consult them through the Internet.
Regarding the type of file that can be stored in a digital repository, there are documents in PDF or other formats, as well as a variety of other files such as photos, images, videos, animations, etc.
In this case, for the COPAGETEC project, it was organized in different Internet hosting because some are more specialized in certain aspects than others.
The main idea of this section is to provide the project source files such as: Modeling, electrical plans, electronic circuits, sensors, firmware, data acquisition during experiments, equations, among others.